Move the tests too
